+++ packages/renderer/src/provider/provider.ts
@@ -58,6 +58,8 @@ import { type ProvideAdapter } from './defaults';
 // 版本信息
 import { version } from '../version';
 
+import { createMenus } from '../hooks';
+
 export const providerKey: InjectionKey<Provider> = Symbol('Provider');
 
 export interface ProviderOptions {
@@ -263,7 +265,7 @@ export class Provider extends Base {
 
         for (const url of urls) {
           if (isCSSUrl(url)) {
-            await loadCss(libraryName, urlUtils.append(url, { v: version }));
+            await loadCss(url, urlUtils.append(url, { v: version }));
           }
           if (isJSUrl(url)) {
             await loadScript(urlUtils.append(url, { v: version }));
@@ -436,6 +438,9 @@ export class Provider extends Base {
       }
     };
     return finder(id, pages) || null;
+  }
+  getMenus(name: string = 'page', prefix: string = '') {
+    return createMenus(prefix, name, this.project?.pages || []);
   }
   getHomepage(): PageFile | null {
     const { homepage } = this.project || {};