+++ docs/CHANGELOG.md
@@ -3,6 +3,68 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.47](https://gitee.com/newgateway/vtj/compare/@vtj/docs@0.12.46...@vtj/docs@0.12.47) (2025-06-24)
+
+**Note:** Version bump only for package @vtj/docs
+
+
+
+
+
+## [0.12.46](https://gitee.com/newgateway/vtj/compare/@vtj/docs@0.12.45...@vtj/docs@0.12.46) (2025-06-23)
+
+**Note:** Version bump only for package @vtj/docs
+
+
+
+
+
+## [0.12.45](https://gitee.com/newgateway/vtj/compare/@vtj/docs@0.12.44...@vtj/docs@0.12.45) (2025-06-23)
+
+**Note:** Version bump only for package @vtj/docs
+
+
+
+
+
+## [0.12.44](https://gitee.com/newgateway/vtj/compare/@vtj/docs@0.12.43...@vtj/docs@0.12.44) (2025-06-20)
+
+
+### Bug Fixes
+
+* 🐛 迁移常量定义 ([d448277](https://gitee.com/newgateway/vtj/commits/d4482775f9626178c0878d42318a89dad0d32f48))
+
+
+
+
+
+## [0.12.43](https://gitee.com/newgateway/vtj/compare/@vtj/docs@0.12.42...@vtj/docs@0.12.43) (2025-06-17)
+
+**Note:** Version bump only for package @vtj/docs
+
+
+
+
+
+## [0.12.42](https://gitee.com/newgateway/vtj/compare/@vtj/docs@0.12.41...@vtj/docs@0.12.42) (2025-06-17)
+
+**Note:** Version bump only for package @vtj/docs
+
+
+
+
+
+## [0.12.41](https://gitee.com/newgateway/vtj/compare/@vtj/docs@0.12.40...@vtj/docs@0.12.41) (2025-06-16)
+
+
+### Features
+
+* ✨ docs wiki ([c00e045](https://gitee.com/newgateway/vtj/commits/c00e0453b01462cc2a6088196aa122a9ead9f751))
+
+
+
+
+
 ## [0.12.40](https://gitee.com/newgateway/vtj/compare/@vtj/docs@0.12.39...@vtj/docs@0.12.40) (2025-06-12)
 
 **Note:** Version bump only for package @vtj/docs