+++ platforms/pro/CHANGELOG.md
@@ -3,6 +3,69 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.47](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.12.46...@vtj/pro@0.12.47) (2025-06-24)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
+## [0.12.46](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.12.45...@vtj/pro@0.12.46) (2025-06-23)
+
+
+### Bug Fixes
+
+* 🐛 uniapp预览跳转链接 ([1d54eaa](https://gitee.com/newgateway/vtj/commits/1d54eaa31274b6b0e192d2f862fea284b71242a6))
+
+
+
+
+
+## [0.12.45](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.12.44...@vtj/pro@0.12.45) (2025-06-23)
+
+
+### Bug Fixes
+
+* 🐛 auth add baseurl ([1f5d9a5](https://gitee.com/newgateway/vtj/commits/1f5d9a538fdd0f5244ba4de0c65a1d78734ab399))
+* 🐛 pro extension ([26b9c66](https://gitee.com/newgateway/vtj/commits/26b9c66eb32b8196413080a327be57a8b039b943))
+
+
+
+
+
+## [0.12.44](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.12.43...@vtj/pro@0.12.44) (2025-06-20)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
+## [0.12.43](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.12.42...@vtj/pro@0.12.43) (2025-06-17)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
+## [0.12.42](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.12.41...@vtj/pro@0.12.42) (2025-06-17)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
+## [0.12.41](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.12.40...@vtj/pro@0.12.41) (2025-06-16)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
 ## [0.12.40](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.12.39...@vtj/pro@0.12.40) (2025-06-12)
 
 **Note:** Version bump only for package @vtj/pro